+++ reservamentor-api/src/main/java/com/reservamentor/integration/payment/paypal/service/Impl/PayPalServiceImpl.java
@@ -1,0 +1,113 @@
+package com.reservamentor.integration.payment.paypal.service.Impl;
+
+
+import com.reservamentor.exception.ResourceNotFoundException;
+import com.reservamentor.integration.payment.paypal.dto.*;
+import com.reservamentor.integration.payment.paypal.service.PayPalService;
+import com.reservamentor.model.entity.Purchase;
+import com.reservamentor.repository.PurchaseRepository;
+import jakarta.annotation.PostConstruct;
+import lombok.NonNull;
+import lombok.RequiredArgsConstructor;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.http.HttpHeaders;
+import org.springframework.http.MediaType;
+import org.springframework.stereotype.Service;
+import org.springframework.util.LinkedMultiValueMap;
+import org.springframework.util.MultiValueMap;
+import org.springframework.web.client.RestClient;
+
+import java.util.Base64;
+import java.util.Collections;
+import java.util.Objects;
+
+@RequiredArgsConstructor
+@Service
+public class PayPalServiceImpl implements PayPalService {
+
+    @Value("${paypal.client-id}")
+    private String clientId;
+
+    @Value("${paypal.client-secret}")
+    private String clientSecret;
+
+    @Value("${paypal.api-base}")
+    private String apiBase;
+
+    @NonNull
+    private PurchaseRepository purchaseRepository;
+    private RestClient paypalClient;
+
+    @PostConstruct
+    public void init(){
+        paypalClient = RestClient
+                .builder()
+                .baseUrl(apiBase)
+                .build();
+    }
+
+    public String getAccessToken(){
+        MultiValueMap<String, String> body = new LinkedMultiValueMap<>();
+        body.add("grant_type", "client_credentials");
+
+        //Realización de Solicitudes:enviando una solicitud POST a la API de PayPal para obtener un token de acceso:
+        return Objects.requireNonNull(
+                        paypalClient.post()
+                                .uri("/v1/oauth2/token")
+                                .contentType(MediaType.APPLICATION_FORM_URLENCODED)
+                                .header(HttpHeaders.AUTHORIZATION, "Basic " + Base64.getEncoder()
+                                        .encodeToString((clientId + ":" + clientSecret).getBytes()))
+                                .body(body)
+                                .retrieve()
+                                .toEntity(TokenResponse.class)
+                                .getBody())
+                .getAccessToken();
+    }
+
+    public OrderResponse createOrder(Integer purchaseId, String returnUrl, String cancelUrl){
+        Purchase purchase= purchaseRepository.findById(purchaseId)
+                .orElseThrow(ResourceNotFoundException::new);
+
+        //Construcción de la solicitud de Pedido de Pago
+        OrderRequest orderRequest = new OrderRequest();
+        orderRequest.setIntent("CAPTURE");
+
+        Amount amount = new Amount();
+        amount.setCurrencyCode("USD");
+        amount.setValue(purchase.getTotal().toString());
+
+        PurchaseUnit purchaseUnit = new PurchaseUnit();
+        purchaseUnit.setReferenceId(purchase.getId().toString());
+        purchaseUnit.setAmount(amount);
+
+        orderRequest.setPurchaseUnits(Collections.singletonList(purchaseUnit));
+
+        ApplicationContext applicationContext = ApplicationContext.builder()
+                .brandName("ReservaMentor")
+                .returnURL(returnUrl)
+                .cancelURL(cancelUrl)
+                .build();
+
+        orderRequest.setApplicationContext(applicationContext);
+
+        return paypalClient.post()
+                .uri("/v2/checkout/orders")
+                .contentType(MediaType.APPLICATION_JSON)
+                .header(HttpHeaders.AUTHORIZATION, "Bearer " + getAccessToken())
+                .body(orderRequest)
+                .retrieve()
+                .toEntity(OrderResponse.class)
+                .getBody();
+    }
+
+
+    public OrderCaptureResponse captureOrder(String orderId){
+        return paypalClient.post()
+                .uri("/v2/checkout/orders/{order_id}/capture", orderId)
+                .contentType(MediaType.APPLICATION_JSON)
+                .header(HttpHeaders.AUTHORIZATION, "Bearer " + getAccessToken())
+                .retrieve()
+                .toEntity(OrderCaptureResponse.class)
+                .getBody();
+    }
+}
