+++ reservamentor-api/src/main/java/com/reservamentor/model/entity/Purchase.java
@@ -3,27 +3,33 @@ package com.reservamentor.model.entity;
 import com.reservamentor.model.entity.enums.PaymentStatus;
 import jakarta.persistence.*;
 import lombok.Data;
+import lombok.Getter;
+import lombok.Setter;
+import org.springframework.web.service.annotation.GetExchange;
 
 import java.time.LocalDateTime;
 import java.util.List;
 
-@Data
+@Getter
+@Setter
 @Entity
-@Table(name = "purchases")
+@Table(name = "purchase")
 public class Purchase {
     @Id
+    @Column(name = "purchaseid", nullable = false)
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Integer id;
+
     private Float total;
+
     private LocalDateTime createdAt;
 
     @Enumerated(EnumType.STRING)
     @Column(name = "payment_status")
     private PaymentStatus paymentStatus;
 
-    @ManyToOne
-    @JoinColumn(name = "user_id", referencedColumnName = "id"
-            , foreignKey = @ForeignKey(name = "FK_purchase_user"))
+    @ManyToOne(fetch = FetchType.EAGER, optional = false)
+    @JoinColumn(name = "usuarioid")
     private Usuario user;
 
     @OneToMany(mappedBy = "purchase", cascade = CascadeType.ALL)
