+++ reservamentor-api/src/main/java/com/reservamentor/repository/TurnoRepository.java
@@ -1,11 +1,9 @@
 package com.reservamentor.repository;
 
-
-import com.reservamentor.model.entity.SesionMentoria;
+import com.reservamentor.model.entity.Turno;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SesionMentoriaRepository extends JpaRepository<SesionMentoria, Integer> {
-
+public interface TurnoRepository extends JpaRepository<Turno, Integer> {
 }
