+++ reservamentor-api/src/main/java/com/reservamentor/repository/UsuarioRepository.java
@@ -2,10 +2,11 @@ package com.reservamentor.repository;
 
 import com.reservamentor.model.entity.Usuario;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
 
 import java.util.Optional;
 
-
+@Repository
 public interface UsuarioRepository extends JpaRepository<Usuario, Integer> {
     boolean existsByCorreo(String correo);
