+++ reservamentor-api/src/main/resources/application-prod.properties
@@ -35,3 +35,5 @@ reservaMentor.openapi.prod-url = http://localhost:8080/api/v1
 
 jwt.secret=chLhMF9w3mwDutysbQxsX8x4CGwZef4mayTGSmbAG2BUsXbYFKvXrVfnPCa62PJxp9TuHxx4PQAS2yWizardDy53j8Uj2wb2AQ3nK8VLg7tUx9HCzHATEp
 jwt.validity-in-seconds=2592000
+
+paypal.client-id="OwO"
