+++ reservamentor-api/src/main/resources/application.properties
@@ -20,7 +20,7 @@ spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.PostgreSQLDialect
 
 #Durante una etapa de desarollo se requiere que este en update
 #Durante produccion se cambia a none
-spring.jpa.hibernate.ddl-auto=update
+spring.jpa.hibernate.ddl-auto=create-drop
 spring.jpa.show-sql=true
 spring.jpa.properties.hibernate.format_sql=true
 
@@ -32,3 +32,7 @@ reservaMentor.openapi.prod-url = http://localhost:8080/api/v1
 
 jwt.secret=chLhMF9w3mwDutysbQxsX8x4CGwZef4mayTGSmbAG2BUsXbYFKvXrVfnPCa62PJxp9TuHxx4PQAS2yWizardDy53j8Uj2wb2AQ3nK8VLg7tUx9HCzHATEp
 jwt.validity-in-seconds=2592000
+
+paypal.client-id=TU-CLIENT-ID
+paypal.client-secret=TU-CLIENT-SECRET
+paypal.api-base= https://api-m.sandbox.paypal.com
